renamed some namespaces to fit the structure
