+++ Looms.PoS/Looms.PoS.Application/Features/Product/Commands/CreateProduct/CreateProductCommandHandler.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using System.Runtime.InteropServices;
 
 namespace Looms.PoS.Application.Features.Product.Commands.CreateProduct;
-//TODO: include ProductVariation and ProductStock repos
+
 public class CreateProductCommandHandler : IRequestHandler<CreateProductCommand, IActionResult>
 {
     private readonly IProductsRepository _productsRepository;
