+++ Looms.PoS/Looms.PoS.Application/Interfaces/ModelsResolvers/IProductVariationModelsResolver.cs
@@ -1,5 +1,5 @@
-﻿using Looms.PoS.Application.Models.Requests.Product;
-using Looms.PoS.Application.Models.Responses.Product;
+﻿using Looms.PoS.Application.Models.Requests.ProductVariation;
+using Looms.PoS.Application.Models.Responses.ProductVariation;
 using Looms.PoS.Domain.Daos;
 
 namespace Looms.PoS.Application.Interfaces.ModelsResolvers;
