+++ Looms.PoS/Looms.PoS.Application/Mappings/ModelsResolvers/ProductVariationModelsResolver.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using Looms.PoS.Application.Interfaces.ModelsResolvers;
-using Looms.PoS.Application.Models.Requests.Product;
-using Looms.PoS.Application.Models.Responses.Product;
+using Looms.PoS.Application.Models.Requests.ProductVariation;
+using Looms.PoS.Application.Models.Responses.ProductVariation;
 using Looms.PoS.Domain.Daos;
 
 namespace Looms.PoS.Application.Mappings.ModelsResolvers;
