+++ Looms.PoS/Looms.PoS.Application/Models/Requests/Product/UpdateProductVariationRequest.cs
@@ -1,8 +1,7 @@
-namespace Looms.PoS.Application.Models.Requests.Product;
+namespace Looms.PoS.Application.Models.Requests.ProductVariation;
 
 public record UpdateProductVariationRequest
 {
-    //TODO: is this the way to link it to Product?
     public Guid ProductId { get; init; }
     public String Name { get; init; } = string.Empty;
     public String Category { get; init; } = string.Empty;
