--- Looms.PoS/Looms.PoS.Domain/Interfaces/IProductStockRepository.cs
@@ -1,11 +1,0 @@
-using Looms.PoS.Domain.Daos;
-
-namespace Looms.PoS.Domain.Interfaces;
-
-public interface IProductStockRepository
-{
-    Task<ProductStockDao> CreateAsync(ProductStockDao productStockDao);
-    Task<IEnumerable<ProductStockDao>> GetAllAsync();
-    Task<ProductStockDao> GetAsync(Guid id);
-    Task<ProductStockDao> UpdateAsync(ProductStockDao productStockDao);
-}
